(chore) refactor vacancy validations for reuse by Vacancy model to ensure no vacancy can be created without the required attributes
Fix specs affected by enforced validations
