+++ .rubocop_todo.yml
@@ -1,17 +1,19 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2017-10-17 12:07:20 +0100 using RuboCop version 0.48.1.
+# on 2018-03-23 19:42:47 +0000 using RuboCop version 0.48.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
 
-# Offense count: 2
-# Cop supports --auto-correct.
-Performance/TimesMap:
+# Offense count: 1
+Lint/AmbiguousOperator:
   Exclude:
-    - 'spec/features/users_can_view_vacancies_spec.rb'
-    - 'spec/presenters/vacancies_presenter_spec.rb'
+    - 'app/form_models/vacancy_form.rb'
+
+# Offense count: 1
+Metrics/AbcSize:
+  Max: 25
 
 # Offense count: 1
 Rails/FilePath:
@@ -19,7 +21,8 @@ Rails/FilePath:
     - 'spec/rails_helper.rb'
 
 # Offense count: 1
-# Cop supports --auto-correct.
-Style/RescueModifier:
+# Configuration parameters: Blacklist.
+# Blacklist: decrement!, decrement_counter, increment!, increment_counter, toggle!, touch, update_all, update_attribute, update_column, update_columns, update_counters
+Rails/SkipsModelValidations:
   Exclude:
-    - 'app/models/vacancy.rb'
+    - 'app/services/publish_vacancy.rb'
