+++ app/controllers/hiring_staff/vacancies/application_details_controller.rb
@@ -1,16 +1,12 @@
 class HiringStaff::Vacancies::ApplicationDetailsController < HiringStaff::Vacancies::ApplicationController
-  before_action :school
+  before_action :school, :redirect_unless_vacancy_session_id
 
   def new
-    redirect_to job_specification_school_vacancy_path(school_id: @school.id) unless session_vacancy_id
-
     @application_details_form = ApplicationDetailsForm.new(session[:vacancy_attributes])
     @application_details_form.valid? if ['step_3', 'review'].include?(session[:current_step])
   end
 
   def create
-    redirect_to job_specification_school_vacancy_path(school_id: @school.id) unless session_vacancy_id
-
     @application_details_form = ApplicationDetailsForm.new(application_details_form)
     store_vacancy_attributes(@application_details_form.vacancy.attributes.compact!)
