+++ app/models/concerns/vacancy_candidate_specification_validations.rb
@@ -1,0 +1,7 @@
+module VacancyCandidateSpecificationValidations
+  extend ActiveSupport::Concern
+
+  included do
+    validates :essential_requirements, presence: true
+  end
+end
