+++ spec/features/job_seekers_can_view_a_vacancy_spec.rb
@@ -23,7 +23,9 @@ RSpec.feature 'Viewing a single published vacancy' do
 
   scenario 'Expired vacancies display a warning message' do
     current_vacancy = create(:vacancy)
-    expired_vacancy = create(:vacancy, :expired)
+    expired_vacancy = build(:vacancy, :expired)
+    expired_vacancy.send :set_slug
+    expired_vacancy.save(validate: false)
 
     visit vacancy_path(current_vacancy)
     expect(page).to have_no_content('This vacancy has expired')
